Reorganize celery image
