+++ celery_worker/requirements.txt
@@ -16,3 +16,4 @@ pybids==0.8.0
 duecredit==0.6.1
 SQLAlchemy==1.3.0
 Flask-Caching==1.3.3
+flask_jwt
